refactor initialize keymanager
