+++ validator/accounts/accounts_enable_disable_test.go
@@ -64,7 +64,7 @@ func TestDisableAccounts_Noninteractive(t *testing.T) {
 	// We attempt to disable the accounts specified.
 	require.NoError(t, DisableAccountsCli(cliCtx))
 
-	keymanager, err := w.InitializeKeymanager(cliCtx.Context, &iface.InitializeKeymanagerConfig{SkipMnemonicConfirm: false})
+	keymanager, err := w.InitializeKeymanager(cliCtx.Context)
 	require.NoError(t, err)
 	remainingAccounts, err := keymanager.FetchValidatingPublicKeys(cliCtx.Context)
 	require.NoError(t, err)
@@ -120,7 +120,7 @@ func TestEnableAccounts_Noninteractive(t *testing.T) {
 	// We attempt to disable the accounts specified.
 	require.NoError(t, DisableAccountsCli(cliCtx))
 
-	km, err := w.InitializeKeymanager(cliCtx.Context, &iface.InitializeKeymanagerConfig{SkipMnemonicConfirm: false})
+	km, err := w.InitializeKeymanager(cliCtx.Context)
 	require.NoError(t, err)
 	remainingAccounts, err := km.FetchValidatingPublicKeys(cliCtx.Context)
 	require.NoError(t, err)
@@ -129,7 +129,7 @@ func TestEnableAccounts_Noninteractive(t *testing.T) {
 	// We attempt to enable the accounts specified.
 	require.NoError(t, EnableAccountsCli(cliCtx))
 
-	km, err = w.InitializeKeymanager(cliCtx.Context, &iface.InitializeKeymanagerConfig{SkipMnemonicConfirm: false})
+	km, err = w.InitializeKeymanager(cliCtx.Context)
 	require.NoError(t, err)
 	remainingAccounts, err = km.FetchValidatingPublicKeys(cliCtx.Context)
 	require.NoError(t, err)
