+++ validator/accounts/iface/wallet.go
@@ -2,7 +2,6 @@ package iface
 
 import (
 	"context"
-	"io"
 
 	"github.com/prysmaticlabs/prysm/validator/keymanager"
 )
@@ -15,16 +14,9 @@ type Wallet interface {
 	AccountsDir() string
 	Password() string
 	// Read methods for important wallet and accounts-related files.
-	ReadEncryptedSeedFromDisk(ctx context.Context) (io.ReadCloser, error)
 	ReadFileAtPath(ctx context.Context, filePath string, fileName string) ([]byte, error)
 	// Write methods to persist important wallet and accounts-related files to disk.
 	WriteFileAtPath(ctx context.Context, pathName string, fileName string, data []byte) error
-	WriteEncryptedSeedToDisk(ctx context.Context, encoded []byte) error
 	// Method for initializing a new keymanager.
-	InitializeKeymanager(ctx context.Context, cfg *InitializeKeymanagerConfig) (keymanager.IKeymanager, error)
-}
-
-type InitializeKeymanagerConfig struct {
-	SkipMnemonicConfirm bool
-	Mnemonic25thWord    string
+	InitializeKeymanager(ctx context.Context) (keymanager.IKeymanager, error)
 }
