+++ validator/accounts/wallet/BUILD.bazel
@@ -9,7 +9,6 @@ go_library(
     deps = [
         "//shared/fileutil:go_default_library",
         "//shared/promptutil:go_default_library",
-        "//validator/accounts/iface:go_default_library",
         "//validator/accounts/prompt:go_default_library",
         "//validator/flags:go_default_library",
         "//validator/keymanager:go_default_library",
